requirements.txt
fake-bpy-module-latest
numpy
